requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.22.2
idna==3.6
imageio==2.34.0
importlib-resources==6.4.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
MyQR==2.3.1
netaddr==1.2.1
netifaces==0.11.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.2
pydeck==0.8.1b0
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
robomaster==0.1.1.68
rpds-py==0.18.0
safetensors==0.4.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typing-extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.43
urllib3==2.2.1
watchdog==4.0.0
zipp==3.18.1
